added 3 colored label

+++ ResMon/Class/Sensors.vb
@@ -676,7 +676,7 @@ Public Class MainboardSensor
         End Try
     End Function
 
-    Public Function RawFanSpeed(Optional fan As Integer = 1) As Single
+    Public Function RawFanSpeed(Optional fan As Integer = 1) As Integer
         Try
             Return CInt(Math.Ceiling(Computer.Hardware.Where(Function(x) x.HardwareType = HardwareType.Mainboard).FirstOrDefault.SubHardware.FirstOrDefault.Sensors.Where(Function(x) x.SensorType = SensorType.Fan AndAlso x.Name = $"Fan #{fan}").FirstOrDefault.Value.GetValueOrDefault))
         Catch ex As Exception
